+++ FLASH6/flash6.js
@@ -20249,7 +20249,7 @@ function requestMobileMockup3dMesh(){
         fw_program:"Altis_Intelligent3_firmware1",
         fw_board:"Altis_Intelligent3_b3",
         fw_protocol:"Flash6-Intelligent-b3",
-        fw_build:"v6 b6"
+        fw_build:"v6 b7"
       };
     }